requirements.txt
tk
numpy
Pillow
opencv-python
scikit-image
PyQt5
matplotlib